lcoa/requirement.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0
PyJWT==2.8.0
PyMySQL==1.0.2
cryptography==41.0.3